product

+++ src/main/java/com/sakda/java/school/phoneshop_night/repository/ModelRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.sakda.java.school.phoneshop_night.entity.Brand;
 import com.sakda.java.school.phoneshop_night.entity.Model;
 
-public interface ModelRepository extends JpaRepository<Model, Integer>{
-	List<Model> findByBrandId(Integer brandId);
+public interface ModelRepository extends JpaRepository<Model, Long>{
+	List<Model> findByBrandId(Long brandId);
 }
