+++ src/main/java/com/sakda/java/school/phoneshop_night/service/BrandService.java
@@ -9,8 +9,8 @@ import com.sakda.java.school.phoneshop_night.entity.Brand;
 
 public interface BrandService {
 	Brand createBrand(Brand brand);
-	Brand getById(Integer id);
-	Brand update(Integer id, Brand brandUpdate);
+	Brand getById(Long id);
+	Brand update(Long id, Brand brandUpdate);
 	List<Brand> getBrands(String name);
 	//List<Brand> getBrands(Map<String, String> params);
 	Page<Brand> getBrands(Map<String, String> params);
